requirements.txt
agentql==0.5.3
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
jiter==0.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
mss==9.0.1
numpy==2.0.1
openai==1.40.3
opencv-python==4.10.0.84
packaging==24.1
playwright==1.45.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rich==13.7.1
rubicon-objc==0.4.9
setuptools==72.1.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
sqlparse==0.5.1
starlette==0.37.2
tf-playwright-stealth==0.0.1
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wheel==0.44.0
